LPD-54760 Move tests to correct folder
